# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MercadoZetta API",
    "version": "1.0.0",
    "description": "HTTP contract for the current white-label marketplace API."
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log in",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "email": "seller@example.com",
                "password": "password123"
              },
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated session; also sets access, refresh, and CSRF cookies",
            "content": {
              "application/json": {
                "example": {
                  "user": {
                    "_id": "507f1f77bcf86cd799439011",
                    "tenantId": "mercadozetta",
                    "email": "seller@example.com",
                    "username": "seller",
                    "telephone": "+55 48 99999-0000"
                  },
                  "session": {
                    "id": "507f1f77bcf86cd799439012",
                    "lastUsedAt": "2026-07-15T12:00:00.000Z",
                    "expiresAt": "2026-07-22T12:00:00.000Z",
                    "absoluteExpiresAt": "2026-08-14T12:00:00.000Z"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "session": {
                      "$ref": "#/components/schemas/Session"
                    }
                  },
                  "required": [
                    "user",
                    "session"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials/token",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid authorization token",
                  "code": "INVALID_AUTH_TOKEN"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/session": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Restore the current cookie session",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          }
        ],
        "responses": {
          "200": {
            "description": "Current user and session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "session": {
                      "$ref": "#/components/schemas/Session"
                    }
                  },
                  "required": [
                    "user",
                    "session"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials/token",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid authorization token",
                  "code": "INVALID_AUTH_TOKEN"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Rotate the current refresh token",
        "security": [
          {
            "refreshCookie": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "schema": {
              "type": "string",
              "description": "Session-bound double-submit proof required for cookie-authenticated mutations.",
              "example": "nonce.signature"
            },
            "required": true,
            "description": "Session-bound double-submit proof required for cookie-authenticated mutations."
          }
        ],
        "responses": {
          "204": {
            "description": "Session rotated and cookies replaced"
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials/token",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid authorization token",
                  "code": "INVALID_AUTH_TOKEN"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A concurrent request already rotated the token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/sessions": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "List active sessions for the current tenant/user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          }
        ],
        "responses": {
          "200": {
            "description": "Active sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    }
                  },
                  "required": [
                    "sessions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials/token",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid authorization token",
                  "code": "INVALID_AUTH_TOKEN"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/sessions/{sessionId}": {
      "delete": {
        "tags": [
          "Authentication"
        ],
        "summary": "Revoke one owned session",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439011"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "schema": {
              "type": "string",
              "description": "Session-bound double-submit proof required for cookie-authenticated mutations.",
              "example": "nonce.signature"
            },
            "required": true,
            "description": "Session-bound double-submit proof required for cookie-authenticated mutations."
          }
        ],
        "responses": {
          "204": {
            "description": "Session revoked"
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials/token",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid authorization token",
                  "code": "INVALID_AUTH_TOKEN"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Session not found",
                  "code": "SESSION_NOT_FOUND"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout/current": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Revoke the current cookie session",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "schema": {
              "type": "string",
              "description": "Session-bound double-submit proof required for cookie-authenticated mutations.",
              "example": "nonce.signature"
            },
            "required": true,
            "description": "Session-bound double-submit proof required for cookie-authenticated mutations."
          }
        ],
        "responses": {
          "204": {
            "description": "Current session revoked"
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials/token",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid authorization token",
                  "code": "INVALID_AUTH_TOKEN"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log out and invalidate existing access tokens",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "schema": {
              "type": "string",
              "description": "Session-bound double-submit proof required for cookie-authenticated mutations.",
              "example": "nonce.signature"
            },
            "required": true,
            "description": "Session-bound double-submit proof required for cookie-authenticated mutations."
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out"
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials/token",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid authorization token",
                  "code": "INVALID_AUTH_TOKEN"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create an account",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "email": "seller@example.com",
                "password": "password123",
                "username": "seller",
                "telephone": "+55 48 99999-0000"
              },
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "example": {
                  "newUser": {
                    "_id": "507f1f77bcf86cd799439011",
                    "tenantId": "mercadozetta",
                    "email": "seller@example.com",
                    "username": "seller",
                    "telephone": "+55 48 99999-0000"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "newUser": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": [
                    "newUser"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a seller profile",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "$ref": "#/components/schemas/SellerId"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          }
        ],
        "responses": {
          "200": {
            "description": "Public seller profile",
            "content": {
              "application/json": {
                "example": {
                  "_id": "507f1f77bcf86cd799439011",
                  "tenantId": "mercadozetta",
                  "email": "seller@example.com",
                  "username": "seller",
                  "telephone": "+55 48 99999-0000",
                  "storeName": "seller store"
                },
                "schema": {
                  "$ref": "#/components/schemas/SellerProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Seller not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Seller not found",
                  "code": "SELLER_NOT_FOUND"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/products/{productId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get a product",
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "schema": {
              "$ref": "#/components/schemas/ProductId"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise.",
              "example": "mercadozetta",
              "type": "string"
            },
            "description": "Tenant slug. Required when TENANT_HEADER_REQUIRED is enabled; defaults to mercadozetta otherwise."
          }
        ],
        "responses": {
          "200": {
            "description": "Product detail",
            "content": {
              "application/json": {
                "example": {
                  "_id": "507f191e810c19729de860ea",
                  "tenantId": "mercadozetta",
                  "name": "mechanical keyboard",
                  "description": "Compact keyboard",
                  "category": "electronics",
                  "subcategory": "keyboards",
                  "inventory": 5,
                  "image": "https://example.com/keyboard.jpg",
                  "status": "active",
                  "seller": "507f1f77bcf86cd799439011"
                },
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or tenant",
            "content": {
              "application/json": {
                "example": {
                  "error": "Invalid tenant",
                  "code": "INVALID_TENANT"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Product not found",
                  "code": "PRODUCT_NOT_FOUND"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "description": "Credentials used to create a tenant-bound session.",
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "SellerId": {
        "type": "string",
        "example": "507f1f77bcf86cd799439011"
      },
...
      "ProductId": {
        "example": "507f191e810c19729de860ea",
        "type": "string"
      },
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "tenantId": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "username": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "_id",
          "tenantId",
          "email",
          "username",
          "telephone"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error",
          "code"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-mz_at"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-mz_rt"
      }
    }
  }
}